+++ src/Control/Eff/LogWriter/Async.hs
@@ -41,7 +41,7 @@ withAsyncLogging
   => LogWriter
   -> len -- ^ Size of the log message input queue. If the queue is full, message
          -- are dropped silently.
-  -> Text -- ^ The default application name to put into the 'logEventAppName' field.
+  -> String -- ^ The default application name to put into the 'logEventAppName' field.
   -> Facility -- ^ The default RFC-5424 facility to put into the 'logEventFacility' field.
   -> LogPredicate -- ^ The inital predicate for log messages, there are some pre-defined in "Control.Eff.Log.Message#PredefinedPredicates"
   -> Eff (Logs : LogWriterReader : e) a