+++ src/Control/Eff/LogWriter/File.hs
@@ -38,7 +38,7 @@ import           Control.Monad.Trans.Control    ( MonadBaseControl
 withFileLogging
   :: (Lifted IO e, MonadBaseControl IO (Eff e), HasCallStack)
   => FilePath -- ^ Path to the log-file.
-  -> Text -- ^ The default application name to put into the 'logEventAppName' field.
+  -> String -- ^ The default application name to put into the 'logEventAppName' field.
   -> Facility -- ^ The default RFC-5424 facility to put into the 'logEventFacility' field.
   -> LogPredicate -- ^ The inital predicate for log messages, there are some pre-defined in "Control.Eff.Log.Message#PredefinedPredicates"
   -> LogEventPrinter -- ^ The 'LogEvent' render function