+++ src/Control/Eff/LogWriter/UDP.hs
@@ -29,7 +29,7 @@ withUDPLogging
   => (LogEvent -> Text) -- ^ 'LogEvent' rendering function
   -> String -- ^ Hostname or IP
   -> String -- ^ Port e.g. @"514"@
-  -> Text -- ^ The default application name to put into the 'logEventAppName' field.
+  -> String -- ^ The default application name to put into the 'logEventAppName' field.
   -> Facility -- ^ The default RFC-5424 facility to put into the 'logEventFacility' field.
   -> LogPredicate -- ^ The inital predicate for log messages, there are some pre-defined in "Control.Eff.Log.Message#PredefinedPredicates"
   -> Eff (Logs : LogWriterReader : e) a