+++ src/Control/Eff/LogWriter/UnixSocket.hs
@@ -28,7 +28,7 @@ withUnixSocketLogging
   :: (HasCallStack, MonadBaseControl IO (Eff e), Lifted IO e)
   => LogEventReader Text -- ^ 'LogEvent' rendering function
   -> FilePath -- ^ Path to the socket file
-  -> Text -- ^ The default application name to put into the 'logEventAppName' field.
+  -> String -- ^ The default application name to put into the 'logEventAppName' field.
   -> Facility -- ^ The default RFC-5424 facility to put into the 'logEventFacility' field.
   -> LogPredicate -- ^ The inital predicate for log messages, there are some pre-defined in "Control.Eff.Log.Message#PredefinedPredicates"
   -> Eff (Logs : LogWriterReader : e) a